Fix util to use ResourceIdentifier
